リアクション処理分離の準備
