+++ models/discordserver.js
@@ -9,6 +9,7 @@ const { ActionContext } = require('../contexts/actioncontext');
 const { VoiceChat } = require('./voicechat');
 const { AudioRequest } = require('./audiorequest');
 const { UserAction, ActionResult } = require('./useraction');
+const { PagingAction, TeachPagingAction, SoundEffectPagingAction } = require('./useraction');
 
 const CommandDelimiterRegexp = new RegExp('[ 　]+');
 
@@ -167,6 +168,41 @@ class DiscordServer {
         }
 
         return Promise.reject('アクションに対応するResponsiveがない');
+    }
+
+    /**
+     * @param {ActionContext} context
+     * @param {discord.MessageReaction} reaction
+     * @param {Buffer} emoji
+     * @returns {Promise<ActionResult>}
+     */
+    async handleReaction(context, reaction, emoji) {
+        const EMOJI_POINT_LEFT = new Uint8Array([0xf0, 0x9f, 0x91, 0x88]);
+        const EMOJI_POINT_RIGHT = new Uint8Array([0xf0, 0x9f, 0x91, 0x89]);
+        /**
+         * @type {Map<string, PagingAction>}
+         * @readonly
+         */
+        const pagerContents = new Map();
+        pagerContents.set('se-list', SoundEffectPagingAction);
+        pagerContents.set('dictionary', TeachPagingAction);
+
+        const commandName = reaction.message.content.split('\n')[0].split(' ')[0];
+        const page = parseInt(reaction.message.content.split('\n')[0].split(' ')[1], 10);
+
+        const pagerActionClass = pagerContents.get(commandName);
+
+        if (!pagerActionClass) {
+            return Promise.reject('アクションに対応するpagerがない');
+        }
+
+        if (emoji.equals(EMOJI_POINT_LEFT)) {
+            const result = await this.handleAction(context, new pagerActionClass(page - 1));
+            return Promise.resolve(result);
+        } else if (emoji.equals(EMOJI_POINT_RIGHT)) {
+            const result = await this.handleAction(context, new pagerActionClass(page + 1));
+            return Promise.resolve(result);
+        }
     }
 
     /**
